Have the WRO filters also handle the REQUEST, FORWARD, and ERROR dispatches
By default, servlet filters only handle REQUEST. Which means that if an ERROR is handled, the filters don't execute. So in this case, if WRO is used during the processing of the ERROR request, the context won't be set (because the wroContextFilter didn't execute) resulting in a context not set exception.